lab6: complete

+++ labcodes_answer/lab6/kern/syscall/syscall.c
@@ -8,27 +8,27 @@
 #include <clock.h>
 
 static int
-sys_exit(uint32_t arg[]) {
+sys_exit(uint64_t arg[]) {
     int error_code = (int)arg[0];
     return do_exit(error_code);
 }
 
 static int
-sys_fork(uint32_t arg[]) {
+sys_fork(uint64_t arg[]) {
     struct trapframe *tf = current->tf;
     uintptr_t stack = tf->gpr.sp;
     return do_fork(0, stack, tf);
 }
 
 static int
-sys_wait(uint32_t arg[]) {
+sys_wait(uint64_t arg[]) {
     int pid = (int)arg[0];
     int *store = (int *)arg[1];
     return do_wait(pid, store);
 }
 
 static int
-sys_exec(uint32_t arg[]) {
+sys_exec(uint64_t arg[]) {
     const char *name = (const char *)arg[0];
     size_t len = (size_t)arg[1];
     unsigned char *binary = (unsigned char *)arg[2];
@@ -37,47 +37,47 @@ sys_exec(uint32_t arg[]) {
 }
 
 static int
-sys_yield(uint32_t arg[]) {
+sys_yield(uint64_t arg[]) {
     return do_yield();
 }
 
 static int
-sys_kill(uint32_t arg[]) {
+sys_kill(uint64_t arg[]) {
     int pid = (int)arg[0];
     return do_kill(pid);
 }
 
 static int
-sys_getpid(uint32_t arg[]) {
+sys_getpid(uint64_t arg[]) {
     return current->pid;
 }
 
 static int
-sys_putc(uint32_t arg[]) {
+sys_putc(uint64_t arg[]) {
     int c = (int)arg[0];
     cputchar(c);
     return 0;
 }
 
 static int
-sys_pgdir(uint32_t arg[]) {
+sys_pgdir(uint64_t arg[]) {
     print_pgdir();
     return 0;
 }
 
 static int
-sys_gettime(uint32_t arg[]) {
+sys_gettime(uint64_t arg[]) {
     return (int)ticks;
 }
 static int
-sys_lab6_set_priority(uint32_t arg[])
+sys_lab6_set_priority(uint64_t arg[])
 {
-    uint32_t priority = (uint32_t)arg[0];
+    uint64_t priority = (uint64_t)arg[0];
     lab6_set_priority(priority);
     return 0;
 }
 
-static int (*syscalls[])(uint32_t arg[]) = {
+static int (*syscalls[])(uint64_t arg[]) = {
     [SYS_exit]              sys_exit,
     [SYS_fork]              sys_fork,
     [SYS_wait]              sys_wait,
